+++ src/app/auth/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignUpForm from '@/components/SignUpForm';
+import SignUpForm from '@/components/auth/SignUpForm';
 import { ROUTES } from '@/constants/routes';
 import Link from 'next/link';
 import React from 'react';
@@ -7,20 +7,6 @@ import type { Metadata } from 'next';
 export const metadata: Metadata = {
   title: '회원 가입',
   description: '면접 타운에 회원 가입해보세요.',
-  // openGraph: {
-  //   title: '회원 가입',
-  //   description: '면접 타운에 회원 가입해보세요.',
-  //   url: 'https://www.interview-town.com/auth/signin',
-  //   siteName: '면접 타운',
-  //   images: [
-  //     {
-  //       url: '/open-graph-image.png',
-  //       width: 1200,
-  //       height: 630,
-  //       alt: '면접 타운 썸네일 이미지',
-  //     },
-  //   ],
-  // },
 };
 
 export default function SignUpPage() {
